+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Change Log
 
+## [v1.2.0] - 2023-10-11
+
+### Added
+
+- Support custom and built-in module KMS for cloudwatch log group
+  - resources: `data.aws_iam_policy_document.cloudwatch_log_group_kms_policy`, `module.cloudwatch_log_group_kms`
+  - variables: `is_create_default_kms`, `cloudwatch_log_group_kms_key_arn`
+- Validation condition `local.raise_multiple_container_attach_to_alb`
+- New method to create task definition with support multiple container `local.container_task_definitions`
+  - variables: `container`
+- Support for 1 secretManager: N secret
+  - resources: `aws_secretsmanager_secret.this`, `aws_secretsmanager_secret_version.this`, `aws_iam_role_policy.task_execution_role_access_secret`
+
+### Changed
+
+- Update example of simple usage `examples/simple/main.tf`, `examples/simple/versions.tf` and `examples/simple/outputs.tf`
+
+### Removed 
+
+- Non-used module level validation `local.raise_vpc_id_empty`, `local.raise_service_port_empty`, `local.raise_health_check_empty` and `local.raise_alb_listener_arn_empty`
+- Remove all previous method to construct the task definition for ECS
+- Remove all secrets usage 1 key : 1 secret; use 1 secret in JSON form
+  - resources: `aws_secretsmanager_secret.service_secrets`, `aws_secretsmanager_secret_version.service_secrets`, `aws_iam_role_policy.task_execution_secrets`
+- Remove unused variables `is_attach_service_with_lb`, `service_info`, `apm_sidecar_ecr_url`, `apm_config`. `unix_max_connection`, `entry_point` and `command`
+
 ## [v1.1.12] - 2023-01-23
 
 ### Added